test-api-server/requirements.txt
grpc-interceptor==0.15.3
grpcio-tools==1.26.0
grpcio==1.49.1
locust==2.16.1
